Split views.py into separate files
