+++ airlock/views/__init__.py
@@ -1,0 +1,14 @@
+from .auth_views import login, logout
+from .index_views import index
+from .request_views import (
+    request_index,
+    request_reject,
+    request_release_files,
+    request_submit,
+    request_view,
+)
+from .workspace_views import (
+    workspace_add_file_to_request,
+    workspace_index,
+    workspace_view,
+)
