+++ airlock/views/index_views.py
@@ -1,0 +1,8 @@
+from django.template.response import TemplateResponse
+
+from .helpers import login_exempt
+
+
+@login_exempt  # for now
+def index(request):
+    return TemplateResponse(request, "index.html")
